feat: build w2c2 in rust build script

+++ build_w2c2.sh
@@ -3,6 +3,11 @@
 set -e
 
 BUILD_DIR=./w2c2
+BINARY_PATH=$BUILD_DIR/build/w2c2/w2c2
+
+if [ -e $BINARY_PATH ]; then
+    exit 0
+fi
 
 rm -rf $BUILD_DIR
 git clone --recursive https://github.com/turbolent/w2c2 $BUILD_DIR
